Initial commit
VuePress static site using @vuepress/theme-blog to generate and manage
posts. Unfortunately, in order to get a desired style out of the theme,
a few overrides where needed, under blog/.vuepress/styles. This results
in the colour scheme seemingly changing (via a one-second transition),
upon page load. To alleviate this, the theme should be forked and have
the overrides changed directly in the theme.

Note a few tweaks were also made in a few of the
blog/.vuepress/components. This is much less of an issue, as these are
completely overridden by the contents of the local components folder,
while CSS changes are kept in the final output (with the local changes
taking precedence).

This is fine for now (as it is much simpler to override styling).

Next steps:
  - writing content
  - refactor to remove styling overrides