feat: Minor link styling changes
Changes some link styling which had visibility issues on hover and
active styles. The surrounding background caused the text color to be
difficult to read in some cases. This is primarily caused by a slight
difference in how the blog color palette is used within the theme I'm
using and the specific colors in the palette I've chosen to define.

A better fix would be to create and use a proper custom theme.